# blood-api/public/openapi.json
{"openapi":"3.0.0","info":{"title":"iDoc API Reference","version":"v1","description":"iDoc Api secification and documentation.","x-logo":{"url":"","altText":"iDoc API Reference","backgroundColor":""}},"components":{"securitySchemes":{"BearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"schemas":[]},"servers":[{"url":"http:\/\/localhost:8000","description":"Documentation generator server."},{"url":"http:\/\/test.example.com","description":"Test server."}],"paths":{"\/api\/createUser":{"post":{"tags":["User"],"operationId":"Create User","description":"","requestBody":{"description":"","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["first_name","last_name","dob","weight","ethnicity","gender","email","phone","password"],"properties":{"first_name":{"type":"string","example":"uiZPE4lc9lLnZB4X","description":""},"last_name":{"type":"string","example":"HYTl0f3QbTq6dphR","description":""},"dob":{"type":"datetime","example":"svf3SFjLPnyfVsla","description":""},"weight":{"type":"integer","example":16,"description":""},"ethnicity":{"type":"string","example":"TaSNZ0SUykCfIyCs","description":""},"gender":{"type":"string","example":"RRDTOrzLKZn8Bmt2","description":""},"email":{"type":"string","example":"IsccwFphbNfgzkwi","description":""},"phone":{"type":"string","example":"r9KySG5UoBRq8dXN","description":""},"password":{"type":"string","example":"zBX8wf3xRGHy3kP0","description":""}},"example":{"first_name":"uiZPE4lc9lLnZB4X","last_name":"HYTl0f3QbTq6dphR","dob":"svf3SFjLPnyfVsla","weight":16,"ethnicity":"TaSNZ0SUykCfIyCs","gender":"RRDTOrzLKZn8Bmt2","email":"IsccwFphbNfgzkwi","phone":"r9KySG5UoBRq8dXN","password":"zBX8wf3xRGHy3kP0"}}}}},"parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"message":"status","status_code":200,"data":[]}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X POST \"http:\/\/localhost:8000\/api\/createUser\" \n    -d \"first_name\"=\"uiZPE4lc9lLnZB4X\" \\\n    -d \"last_name\"=\"HYTl0f3QbTq6dphR\" \\\n    -d \"dob\"=\"svf3SFjLPnyfVsla\" \\\n    -d \"weight\"=\"16\" \\\n    -d \"ethnicity\"=\"TaSNZ0SUykCfIyCs\" \\\n    -d \"gender\"=\"RRDTOrzLKZn8Bmt2\" \\\n    -d \"email\"=\"IsccwFphbNfgzkwi\" \\\n    -d \"phone\"=\"r9KySG5UoBRq8dXN\" \\\n    -d \"password\"=\"zBX8wf3xRGHy3kP0\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/createUser\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"first_name\": \"uiZPE4lc9lLnZB4X\",\n    \"last_name\": \"HYTl0f3QbTq6dphR\",\n    \"dob\": \"svf3SFjLPnyfVsla\",\n    \"weight\": \"16\",\n    \"ethnicity\": \"TaSNZ0SUykCfIyCs\",\n    \"gender\": \"RRDTOrzLKZn8Bmt2\",\n    \"email\": \"IsccwFphbNfgzkwi\",\n    \"phone\": \"r9KySG5UoBRq8dXN\",\n    \"password\": \"zBX8wf3xRGHy3kP0\",\n})\n\nfetch(url, {\n    method: \"POST\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"first_name\" =>  \"uiZPE4lc9lLnZB4X\",\n    \"last_name\" =>  \"HYTl0f3QbTq6dphR\",\n    \"dob\" =>  \"svf3SFjLPnyfVsla\",\n    \"weight\" =>  \"16\",\n    \"ethnicity\" =>  \"TaSNZ0SUykCfIyCs\",\n    \"gender\" =>  \"RRDTOrzLKZn8Bmt2\",\n    \"email\" =>  \"IsccwFphbNfgzkwi\",\n    \"phone\" =>  \"r9KySG5UoBRq8dXN\",\n    \"password\" =>  \"zBX8wf3xRGHy3kP0\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/createUser\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_setopt($ch, CURLOPT_POSTFIELDS, $body);\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/authenticateUser":{"post":{"tags":["User"],"operationId":"Authenticate User\nCreate a usable token for the logged in user","description":"","requestBody":{"description":"","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string","example":"oChkoE4YLXo2fwVD","description":""},"password":{"type":"string","example":"QCWtnoVg1fInkcEC","description":""}},"example":{"email":"oChkoE4YLXo2fwVD","password":"QCWtnoVg1fInkcEC"}}}}},"parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"message":"status","status_code":200,"data":[]}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X POST \"http:\/\/localhost:8000\/api\/authenticateUser\" \n    -d \"email\"=\"oChkoE4YLXo2fwVD\" \\\n    -d \"password\"=\"QCWtnoVg1fInkcEC\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/authenticateUser\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"email\": \"oChkoE4YLXo2fwVD\",\n    \"password\": \"QCWtnoVg1fInkcEC\",\n})\n\nfetch(url, {\n    method: \"POST\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"email\" =>  \"oChkoE4YLXo2fwVD\",\n    \"password\" =>  \"QCWtnoVg1fInkcEC\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/authenticateUser\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_setopt($ch, CURLOPT_POSTFIELDS, $body);\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/updateUser\/{id}":{"put":{"tags":["User"],"operationId":"Update User","description":"","requestBody":{"description":"","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["first_name","last_name","dob","weight","ethnicity","gender","phone","password"],"properties":{"first_name":{"type":"string","example":"aqrTFr4nKyWdZD8Y","description":""},"last_name":{"type":"string","example":"0CXXxrKuiXBtUgsj","description":""},"dob":{"type":"datetime","example":"hu6D89iUshueE9yI","description":""},"weight":{"type":"integer","example":11,"description":""},"ethnicity":{"type":"string","example":"rYu7UNUlkA4zbFe3","description":""},"gender":{"type":"string","example":"1Utdi0FDnb3HqopQ","description":""},"phone":{"type":"string","example":"PXWaGzZLn2noFEbZ","description":""},"password":{"type":"string","example":"wLMyNHs5xS0UbrRN","description":""}},"example":{"first_name":"aqrTFr4nKyWdZD8Y","last_name":"0CXXxrKuiXBtUgsj","dob":"hu6D89iUshueE9yI","weight":11,"ethnicity":"rYu7UNUlkA4zbFe3","gender":"1Utdi0FDnb3HqopQ","phone":"PXWaGzZLn2noFEbZ","password":"wLMyNHs5xS0UbrRN"}}}}},"parameters":[{"in":"path","name":"id","description":"Users id","required":true,"schema":{"type":"integer","example":14}}],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"message":"status","status_code":200,"data":[]}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X PUT \"http:\/\/localhost:8000\/api\/updateUser\/{id}\" \n    -d \"first_name\"=\"aqrTFr4nKyWdZD8Y\" \\\n    -d \"last_name\"=\"0CXXxrKuiXBtUgsj\" \\\n    -d \"dob\"=\"hu6D89iUshueE9yI\" \\\n    -d \"weight\"=\"11\" \\\n    -d \"ethnicity\"=\"rYu7UNUlkA4zbFe3\" \\\n    -d \"gender\"=\"1Utdi0FDnb3HqopQ\" \\\n    -d \"phone\"=\"PXWaGzZLn2noFEbZ\" \\\n    -d \"password\"=\"wLMyNHs5xS0UbrRN\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/updateUser\/{id}\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"first_name\": \"aqrTFr4nKyWdZD8Y\",\n    \"last_name\": \"0CXXxrKuiXBtUgsj\",\n    \"dob\": \"hu6D89iUshueE9yI\",\n    \"weight\": \"11\",\n    \"ethnicity\": \"rYu7UNUlkA4zbFe3\",\n    \"gender\": \"1Utdi0FDnb3HqopQ\",\n    \"phone\": \"PXWaGzZLn2noFEbZ\",\n    \"password\": \"wLMyNHs5xS0UbrRN\",\n})\n\nfetch(url, {\n    method: \"PUT\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"first_name\" =>  \"aqrTFr4nKyWdZD8Y\",\n    \"last_name\" =>  \"0CXXxrKuiXBtUgsj\",\n    \"dob\" =>  \"hu6D89iUshueE9yI\",\n    \"weight\" =>  \"11\",\n    \"ethnicity\" =>  \"rYu7UNUlkA4zbFe3\",\n    \"gender\" =>  \"1Utdi0FDnb3HqopQ\",\n    \"phone\" =>  \"PXWaGzZLn2noFEbZ\",\n    \"password\" =>  \"wLMyNHs5xS0UbrRN\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/updateUser\/{id}\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"PUT\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/setBloodGroup\/{id}":{"post":{"tags":["User"],"operationId":"Set Blood Group","description":"","requestBody":{"description":"","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["bloodGroup"],"properties":{"bloodGroup":{"type":"string","example":"jRvoHkQrb8U9b8YH","description":""}},"example":{"bloodGroup":"jRvoHkQrb8U9b8YH"}}}}},"parameters":[{"in":"path","name":"id","description":"Users id","required":true,"schema":{"type":"integer","example":17}}],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"message":"status","status_code":200,"data":[]}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X POST \"http:\/\/localhost:8000\/api\/setBloodGroup\/{id}\" \n    -d \"bloodGroup\"=\"jRvoHkQrb8U9b8YH\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/setBloodGroup\/{id}\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"bloodGroup\": \"jRvoHkQrb8U9b8YH\",\n})\n\nfetch(url, {\n    method: \"POST\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"bloodGroup\" =>  \"jRvoHkQrb8U9b8YH\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/setBloodGroup\/{id}\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_setopt($ch, CURLOPT_POSTFIELDS, $body);\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/sendBloodDonorRequest":{"post":{"tags":["Donation"],"operationId":"Send Donation Request","description":"","requestBody":{"description":"","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["recepient_id","max_acceptance_no","metadata"],"properties":{"recepient_id":{"type":"integer","example":11,"description":""},"max_acceptance_no":{"type":"integer","example":11,"description":""},"metadata":{"type":"object","example":null,"description":""}},"example":{"recepient_id":11,"max_acceptance_no":11,"metadata":null}}}}},"parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"message":"status","status_code":200,"data":[]}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X POST \"http:\/\/localhost:8000\/api\/sendBloodDonorRequest\" \n    -d \"recepient_id\"=\"11\" \\\n    -d \"max_acceptance_no\"=\"11\" \\\n    -d \"metadata\"=\"NqtyIBfYguP11GEG\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/sendBloodDonorRequest\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"recepient_id\": \"11\",\n    \"max_acceptance_no\": \"11\",\n    \"metadata\": NqtyIBfYguP11GEG,\n})\n\nfetch(url, {\n    method: \"POST\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"recepient_id\" =>  \"11\",\n    \"max_acceptance_no\" =>  \"11\",\n    \"metadata\" => \"json_decode(NqtyIBfYguP11GEG, true)\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/sendBloodDonorRequest\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_setopt($ch, CURLOPT_POSTFIELDS, $body);\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/acceptBloodDonorRequest":{"post":{"tags":["Donation"],"operationId":"Accept Donation Request","description":"","requestBody":{"description":"","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["recepient_id","request_id","acceptor_id"],"properties":{"recepient_id":{"type":"integer","example":13,"description":""},"request_id":{"type":"integer","example":3,"description":""},"acceptor_id":{"type":"integer","example":15,"description":""}},"example":{"recepient_id":13,"request_id":3,"acceptor_id":15}}}}},"parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"message":"status","status_code":200,"data":[]}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X POST \"http:\/\/localhost:8000\/api\/acceptBloodDonorRequest\" \n    -d \"recepient_id\"=\"13\" \\\n    -d \"request_id\"=\"3\" \\\n    -d \"acceptor_id\"=\"15\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/acceptBloodDonorRequest\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"recepient_id\": \"13\",\n    \"request_id\": \"3\",\n    \"acceptor_id\": \"15\",\n})\n\nfetch(url, {\n    method: \"POST\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"recepient_id\" =>  \"13\",\n    \"request_id\" =>  \"3\",\n    \"acceptor_id\" =>  \"15\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/acceptBloodDonorRequest\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_setopt($ch, CURLOPT_POSTFIELDS, $body);\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/getUserNotifications\/{id}":{"get":{"tags":["general"],"operationId":"Get All User Notifications","description":"","parameters":[{"in":"path","name":"id","description":"Users id","required":true,"schema":{"type":"integer","example":11}}],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"message":"status","status_code":200,"data":[]}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X GET -G \"http:\/\/localhost:8000\/api\/getUserNotifications\/{id}\" \n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/getUserNotifications\/{id}\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"GET\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/getUserNotifications\/{id}\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/getNotificationDetails\/{id}":{"get":{"tags":["general"],"operationId":"Get Notification Details","description":"","parameters":[{"in":"path","name":"id","description":"Users id","required":true,"schema":{"type":"integer","example":9}}],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"message":"status","status_code":200,"data":[]}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X GET -G \"http:\/\/localhost:8000\/api\/getNotificationDetails\/{id}\" \n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/getNotificationDetails\/{id}\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"GET\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/getNotificationDetails\/{id}\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/readNotification\/{id}":{"get":{"tags":["general"],"operationId":"Read Notifications","description":"","parameters":[{"in":"path","name":"id","description":"Users id","required":true,"schema":{"type":"integer","example":4}}],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"message":"status","status_code":200,"data":[]}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X GET -G \"http:\/\/localhost:8000\/api\/readNotification\/{id}\" \n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost:8000\/api\/readNotification\/{id}\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"GET\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost:8000\/api\/readNotification\/{id}\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}}}}